+++ SP_Lab3/src/main/java/ro/uvt/persistence/BooksRepository.java
@@ -1,3 +1,4 @@
+
 package ro.uvt.persistence;
 
 import org.springframework.data.jpa.repository.JpaRepository;
@@ -5,5 +6,5 @@ import org.springframework.stereotype.Repository;
 import ro.uvt.models.Book;
 
 @Repository
-public interface BooksRepository extends JpaRepository<Book, Integer> {
+public interface BooksRepository extends JpaRepository<Book, Long> {
 }
